requirements.txt
nltk~=3.8.1
numpy~=1.26.4
pandas~=2.2.2
matplotlib~=3.8.4
tensorboard~=2.17.0
scikit-learn~=1.4.2
tabulate~=0.9.0
tqdm~=4.66.2
torchmetrics~=1.4.0.post0
transformers~=4.39.3
tokenizers~=0.15.2
regex~=2023.12.25
pyyaml~=6.0.1
pytorch-lightning~=2.2.2